Huge update
I've spent A LOT of time neglecting my dotfiles repo. The reason for
that was me being lazy :x

That said, a lot of my tools and config have changes, as well as my
understanding of my setup, so I've decided to take a careful look and
update these files properly.